[automation/houdini] darol plugin - network refactor
